rearrange code and improve commenting
general cleanup for readability and code layout.

- dataNode handling has been moved into its own file.
- errCore handling has been moved into its own file.
- other files have been rearranged.

no logical changes have occurred.  The current changes are purely
superficial.
